requirements.txt
python-evtx
xmltodict
requests